requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
color==0.1
colorhash==1.0.2
colour==0.1.5
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.15
django-allauth==0.42.0
django-celery-results==1.2.1
django-colorfield==0.3.2
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.1.3
django-extra-fields==0.3
django-filter==2.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-nine==0.2.3
django-randompinfield==0.4.7
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
elasticsearch==7.8.1
elasticsearch-dsl==7.2.1
enum34==1.1.10
gunicorn==20.0.4
idna==2.10
kombu==4.6.11
Markdown==3.2.2
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
python-barcode==0.13.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
serpy==0.3.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
whitenoise==5.2.0
